Adding confirmation to avery action on ec2-start-stop

+++ aws/ec2-start-stop/main.py
@@ -21,18 +21,25 @@ def get_running_instances():
 def ec2_action(instances_ids):
     try:
         action = sys.argv[1]
+        confirmation = input('Are you sure you want to ' + action + ' these instances? [y/n]:' + str(instances_ids))
 
-        if action == 'start':
-            logging.info('Starting instances: ' + str(instances_ids))
-            ec2.start_instances(InstanceIds=instances_ids)
-        elif action == 'stop':
-            logging.info('Stopping instances: ' + str(instances_ids))
-            ec2.stop_instances(InstanceIds=instances_ids)
-        elif action == 'terminate':
-            logging.info('Terminating instances: ' + str(instances_ids))
-            ec2.terminate_instances(InstanceIds=instances_ids)
+        if confirmation == 'y':
+            if action == 'start':
+                logging.info('Starting instances: ' + str(instances_ids))
+                ec2.start_instances(InstanceIds=instances_ids)
+            elif action == 'stop':
+                logging.info('Stopping instances: ' + str(instances_ids))
+                ec2.stop_instances(InstanceIds=instances_ids)
+            elif action == 'terminate':
+                logging.info('Terminating instances: ' + str(instances_ids))
+                ec2.terminate_instances(InstanceIds=instances_ids)
+            else:
+                logging.info('Invalid action: ' + action)
+        elif confirmation == 'n':
+            logging.info('No instances to ' + action)
         else:
-            logging.info('Invalid action: ' + action)
+            logging.info(f"Invalid input: {confirmation}")
+    
     except Exception as e:
         logging.error(f"Unable to perform action {e}")
